requirements.txt
# MCP Data Steward Dependencies
# Core
pydantic==2.5.0
python-dotenv==1.0.0
mcp[cli]==1.7.1

# Database connectors
psycopg2-binary==2.9.9
mysql-connector-python==8.2.0
cx-oracle==8.3.0
sqlalchemy==2.0.23

# AI/ML (optional)
openai==1.3.0
transformers==4.35.0
torch==2.1.0

# Security and validation
sqlparse==0.4.4
sqlalchemy-utils==0.41.1

# File system and storage
boto3==1.34.0
aiofiles==23.2.0

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1